Fix Link header parser for URLs with semicolons
The `>` character is a more canonical delimiter for the end of a URI, and
fixes the original issue of failed parses of URIs containing a valid
semicolon.

The original use of semicolon and backtracking it seems intentional, but
not sure what the original intent or test case might have been.